refactor: move ActionButton component
