+++ web/components/action-buttons/FollowButton.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 import { useRecoilValue } from 'recoil';
 import Modal from '../ui/Modal/Modal';
 import FollowModal from '../modals/Follow/FollowModal';
-import s from './ActionButton.module.scss';
+import s from './ActionButton/ActionButton.module.scss';
 import { clientConfigStateAtom } from '../stores/ClientConfigStore';
 import { ClientConfig } from '../../interfaces/client-config.model';
