+++ web/components/ui/Content/Content.tsx
@@ -26,8 +26,8 @@ import Footer from '../Footer';
 // import ChatContainer from '../../chat/ChatContainer';
 // import { ChatMessage } from '../../../interfaces/chat-message.model';
 // import ChatTextField from '../../chat/ChatTextField/ChatTextField';
-import ActionButtonRow from '../../action-buttons/ActionButtonRow';
-import ActionButton from '../../action-buttons/ActionButton';
+import ActionButtonRow from '../../action-buttons/ActionButtonRow/ActionButtonRow';
+import ActionButton from '../../action-buttons/ActionButton/ActionButton';
 import NotifyReminderPopup from '../NotifyReminderPopup/NotifyReminderPopup';
 import OfflineBanner from '../OfflineBanner/OfflineBanner';
 import { AppStateOptions } from '../../stores/application-state';
